To allow Git for Windows to carry a temporary state of Cygwin (because
we do not want to wait for Cygwin v3.3.5 before releasing Git for
Windows v2.36.1), we merge a cherry-picked version of the patches in
`cygwin-3_3-branch`.

+++ winsup/cygwin/exceptions.cc
@@ -1139,14 +1139,39 @@ ctrl_c_handler (DWORD type)
     }
 
   if (ch_spawn.set_saw_ctrl_c ())
-    return TRUE;
+    {
+      if (myself->process_state & PID_NOTCYGWIN)
+	sigExeced = SIGINT;
+      return TRUE;
+    }
 
   /* We're only the process group leader when we have a valid pinfo structure.
      If we don't have one, then the parent "stub" will handle the signal. */
   if (!pinfo (cygwin_pid (GetCurrentProcessId ())))
     return TRUE;
 
+  if (type == CTRL_C_EVENT && ::cygheap->ctty
+      && !cygheap->ctty->need_console_handler ())
+    /* Ctrl-C is handled in fhandler_console::cons_master_thread(). */
+    return TRUE;
+
   tty_min *t = cygwin_shared->tty.get_cttyp ();
+
+  if (!t)
+    return TRUE;
+
+  /* If process group leader is non-cygwin process or not exist,
+     send signal to myself. */
+  pinfo pi (t->getpgid ());
+  if ((!pi || (pi->process_state & PID_NOTCYGWIN))
+      && (!have_execed || have_execed_cygwin)
+      && t->getpgid () == myself->pgid
+      && type == CTRL_C_EVENT)
+    {
+      t->output_stopped = false;
+      sig_send(myself, SIGINT);
+    }
+
   /* Ignore this if we're not the process group leader since it should be
      handled *by* the process group leader. */
   if (t && (!have_execed || have_execed_cygwin)
@@ -1556,6 +1581,8 @@ dosig:
   if (have_execed)
     {
       sigproc_printf ("terminating captive process");
+      if (::cygheap->ctty)
+	::cygheap->ctty->cleanup_before_exit ();
       TerminateProcess (ch_spawn, sigExeced = si.si_signo);
     }
   /* Dispatch to the appropriate function. */
